requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
